chore: refactoring code
