+++ carsus/io/hdf.py
@@ -8,7 +8,7 @@ import pandas as pd
 import h5py
 from astropy import units as u
 
-from carsus.metadata import MetadataHandler
+from carsus.metadata.metadata import MetadataHandler
 
 def save_to_hdf(df: pd.DataFrame, 
                path: str,
